docker

+++ config/node/docker-entrypoint.sh
@@ -1,10 +1,18 @@
 #!/bin/bash
 
 if [ "$1" = 'npm' ] ; then
-  npm install
-  if [[ ! -f ".env" ]]; then
+  if [ "$APP_ENV" != 'prod' ]; then
+    npm install
+  fi
+
+  if [[ ! -f ".env" ]] && [ "$APP_ENV" != 'prod' ]; then
     cp .env.dist .env
   fi
+
+  if [ "$1" = 'npm' ]  && [ "$2" = 'run' ] && [ "$3" = 'start' ]; then
+     npm run build
+  fi
+
   echo -e "\e[30;48;5;82mergonode frontend is available at http://localhost:${EXPOSED_NODE_PORT} \e[0m"
 fi
